requirements.txt
Django==3.2.4
djangorestframework==3.12.4
mysqlclient==2.0.3
django-mysql==3.12
django-filter
PyJWT==2.1.0
Pillow==8.3.1
django-cors-headers==3.7.0